+++ visvis2/__init__.py
@@ -1,8 +1,8 @@
 # flake8: noqa
 
 from . import utils
-from ._wrappers import BufferWrapper
 
+from .datawrappers import *
 from .objects import *
 from .geometries import *
 from .materials import *
