+++ visvis2/cameras/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa
+
 from ._base import Camera, NDCCamera, ScreenCoordsCamera
 from ._perspective import PerspectiveCamera
