+++ visvis2/datawrappers/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+
+from ._bufferwrapper import BufferWrapper
