+++ visvis2/geometries/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa
+
 from ._base import Geometry
 from ._box import BoxGeometry
