+++ visvis2/geometries/_box.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from .._wrappers import BufferWrapper
+from ..datawrappers import BufferWrapper
 from ._base import Geometry
