+++ visvis2/materials/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
+
 from ._base import Material
 from ._mesh_basic import MeshBasicMaterial
 from ._line import LineStripMaterial
