+++ visvis2/objects/__init__.py
@@ -1,8 +1,7 @@
 # flake8: noqa
+
 from ._base import WorldObject
 from ._group import Group
 from ._scene import Scene
 from ._mesh import Mesh
 from ._line import Line
-
-# from ._line import Line
