+++ visvis2/renderers/svg/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa
+
 from ._svgrenderer import SvgRenderer, registry, register_svg_render_function
 from . import linerender
