+++ visvis2/renderers/wgpu/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
+
 from ._wgpurenderer import stdinfo_uniform_type, WgpuRenderer
 from ._wgpurenderer import registry, register_wgpu_render_function
 from . import meshrender
